F nabilla search train

+++ tkpdtrain/src/main/java/com/tokopedia/tkpdtrain/search/data/TrainScheduleDataStoreFactory.java
@@ -5,12 +5,21 @@ import com.tokopedia.tkpdtrain.common.specification.DbFlowSpecification;
 import com.tokopedia.tkpdtrain.common.specification.Specification;
 import com.tokopedia.tkpdtrain.search.data.entity.ScheduleAvailabilityEntity;
 import com.tokopedia.tkpdtrain.search.data.entity.TrainListSchedulesEntity;
+import com.tokopedia.tkpdtrain.search.data.specification.TrainAvailabilitySpecification;
+import com.tokopedia.tkpdtrain.search.data.specification.TrainScheduleClassFilterSpecification;
+import com.tokopedia.tkpdtrain.search.data.specification.TrainScheduleNameFilterSpecification;
+import com.tokopedia.tkpdtrain.search.data.specification.TrainSchedulePriceFilterSpecification;
+import com.tokopedia.tkpdtrain.search.data.specification.TrainScheduleSortSpecification;
+import com.tokopedia.tkpdtrain.search.data.specification.TrainScheduleSpecification;
+import com.tokopedia.tkpdtrain.search.data.typedef.TrainScheduleTypeDef;
 import com.tokopedia.tkpdtrain.search.domain.FilterParam;
+import com.tokopedia.tkpdtrain.search.domain.FilterSearchData;
 import com.tokopedia.tkpdtrain.search.domain.mapper.AvailabilityKeysMapper;
 import com.tokopedia.tkpdtrain.search.presentation.model.AvailabilityKeySchedule;
 import com.tokopedia.tkpdtrain.search.presentation.model.TrainScheduleViewModel;
 
 import java.util.List;
+import java.util.Map;
 
 import rx.Observable;
 import rx.functions.Func1;
@@ -33,25 +42,31 @@ public class TrainScheduleDataStoreFactory {
         this.cloudDataStore = cloudDataStore;
     }
 
-    public Observable<List<AvailabilityKeySchedule>> getScheduleTrain(final Specification specification) {
-        return dbDataStore.deleteAll().flatMap(new Func1<Boolean, Observable<List<AvailabilityKeySchedule>>>() {
-            @Override
-            public Observable<List<AvailabilityKeySchedule>> call(Boolean isSuccessDelete) {
-                if (isSuccessDelete) {
-                    return getDataFromCloud(specification);
-                } else {
-                    return Observable.empty();
+    public Observable<List<AvailabilityKeySchedule>> getScheduleTrain(final Specification specification,
+                                                                      final int scheduleVariant) {
+        if (scheduleVariant == TrainScheduleTypeDef.RETURN_SCHEDULE) {
+            return getDataFromCloud(specification, scheduleVariant);
+        } else if (scheduleVariant == TrainScheduleTypeDef.DEPARTURE_SCHEDULE) {
+            return dbDataStore.deleteAll().flatMap(new Func1<Boolean, Observable<List<AvailabilityKeySchedule>>>() {
+                @Override
+                public Observable<List<AvailabilityKeySchedule>> call(Boolean isSuccessDelete) {
+                    if (isSuccessDelete) {
+                        return getDataFromCloud(specification, scheduleVariant);
+                    } else {
+                        return Observable.empty();
+                    }
                 }
-            }
-        });
+            });
+        }
+        return null;
     }
 
-    private Observable<List<AvailabilityKeySchedule>> getDataFromCloud(Specification specification) {
+    private Observable<List<AvailabilityKeySchedule>> getDataFromCloud(Specification specification, final int scheduleVariant) {
         return cloudDataStore.getDatasSchedule(specification)
                 .flatMap(new Func1<TrainListSchedulesEntity, Observable<List<AvailabilityKeySchedule>>>() {
                     @Override
                     public Observable<List<AvailabilityKeySchedule>> call(final TrainListSchedulesEntity trainListSchedulesEntity) {
-                        return dbDataStore.insertAll(trainListSchedulesEntity.getTrainSchedules())
+                        return dbDataStore.insertAllData(trainListSchedulesEntity.getTrainSchedules(), scheduleVariant)
                                 .flatMap(new Func1<Boolean, Observable<List<AvailabilityKeySchedule>>>() {
                                     @Override
                                     public Observable<List<AvailabilityKeySchedule>> call(Boolean isSuccessSaveData) {
@@ -67,7 +82,7 @@ public class TrainScheduleDataStoreFactory {
                 });
     }
 
-    public Observable<List<TrainScheduleViewModel>> getAvailabilitySchedule(String idTrain) {
+    public Observable<List<TrainScheduleViewModel>> getAvailabilitySchedule(String idTrain, final int scheduleVariant) {
         return cloudDataStore.getDatasAvailability(idTrain)
                 .flatMap(new Func1<List<ScheduleAvailabilityEntity>, Observable<List<TrainScheduleViewModel>>>() {
                     @Override
@@ -79,7 +94,7 @@ public class TrainScheduleDataStoreFactory {
                                         if (!isSuccessSavedData) {
                                             return Observable.empty();
                                         } else {
-                                            return dbDataStore.getDatas(new TrainAvailabilitySpecification(scheduleAvailabilityEntities));
+                                            return dbDataStore.getDatas(new TrainAvailabilitySpecification(scheduleAvailabilityEntities, scheduleVariant));
                                         }
                                     }
                                 });
@@ -102,4 +117,26 @@ public class TrainScheduleDataStoreFactory {
         return dbDataStore.getDatas(specification);
     }
 
+    public Observable<Integer> getCountSchedule(FilterSearchData filterSearchData) {
+        DbFlowSpecification specification = new TrainSchedulePriceFilterSpecification(filterSearchData.getMinPrice(), filterSearchData.getMaxPrice());
+        if (filterSearchData.getTrainClass() != null && !filterSearchData.getTrainClass().isEmpty()) {
+            specification = new AndDbFlowSpecification(specification,
+                    new TrainScheduleClassFilterSpecification(filterSearchData.getTrainClass()));
+        }
+        if (filterSearchData.getTrainClass() != null && !filterSearchData.getTrains().isEmpty()) {
+            specification = new AndDbFlowSpecification(specification,
+                    new TrainScheduleNameFilterSpecification(filterSearchData.getTrains()));
+        }
+        return dbDataStore.getCount(specification);
+    }
+
+    public Observable<TrainScheduleViewModel> getDetailScheduleById(Specification specification) {
+        return dbDataStore.getData(specification);
+    }
+
+    public Observable<List<TrainScheduleViewModel>> getFilterSearchParamData(Map<String, Object> mapParam, int scheduleVariant) {
+        TrainScheduleSpecification trainScheduleSpecification = new TrainScheduleSpecification(mapParam);
+        trainScheduleSpecification.setScheduleVariant(scheduleVariant);
+        return dbDataStore.getDatas(trainScheduleSpecification);
+    }
 }
